[openwebnet] Changes after review: extract LightAutomHandlersMap class to own file
